+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/tests/TransactionService.scala
@@ -12,7 +12,6 @@ import com.daml.ledger.api.testtool.infrastructure.{LedgerSession, LedgerTestSui
 import com.digitalasset.ledger.api.v1.transaction.{Transaction, TransactionTree}
 import com.digitalasset.ledger.client.binding.Primitive
 import com.digitalasset.ledger.client.binding.Value.encode
-import com.digitalasset.ledger.test_dev.Test.TextContainer
 import com.digitalasset.ledger.test_stable.Test.Agreement._
 import com.digitalasset.ledger.test_stable.Test.AgreementFactory._
 import com.digitalasset.ledger.test_stable.Test.Choice1._
@@ -779,34 +778,6 @@ class TransactionService(session: LedgerSession) extends LedgerTestSuite(session
           "NoReorder",
           fields.flatMap(_.getValue.getRecord.fields).map(_.getValue.getText).zipWithIndex,
           Seq("street" -> 0, "city" -> 1, "state" -> 2, "zip" -> 3))
-      }
-  }
-
-  test(
-    "TXLargeCommand",
-    "Accept huge submissions with a large number of commands",
-    allocate(SingleParty)) {
-    case Participants(Participant(ledger, party)) =>
-      val targetNumberOfSubCommands = 1500
-      for {
-        request <- ledger.submitAndWaitRequest(
-          party,
-          List.fill(targetNumberOfSubCommands)(Dummy(party).create.command): _*)
-        result <- ledger.submitAndWaitForTransaction(request)
-      } yield {
-        val _ = assertLength("LargeCommand", targetNumberOfSubCommands, result.events)
-      }
-  }
-
-  test("TXManyCommands", "Accept many, large commands at once", allocate(SingleParty)) {
-    case Participants(Participant(ledger, party)) =>
-      val targetNumberOfCommands = 500
-      val oneKbOfText = new String(Array.fill(512 /* two bytes each */ )('a'))
-      for {
-        contractIds <- Future.sequence((1 to targetNumberOfCommands).map(_ =>
-          ledger.create(party, TextContainer(party, oneKbOfText))))
-      } yield {
-        val _ = assertLength("ManyCommands", targetNumberOfCommands, contractIds)
       }
   }
